% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artic_sea_ice.R
\name{artic_sea_ice}
\alias{artic_sea_ice}
\title{Download Arctic ice extent}
\usage{
artic_sea_ice(years = NULL, pole = "north", window = 5)
}
\arguments{
\item{years}{}
}
\description{
from here \link\link{https://nsidc.org/arcticseaicenews/charctic-interactive-sea-ice-graph/}
For now only for the north (artic) and with 5-day moving average
}
