% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_au.R
\docType{data}
\name{bls_au}
\alias{bls_au}
\title{Data for the bundesländer of Austria}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.
}
\usage{
bls_au
}
\description{
Data for the bundesländer of Austria
}
\keyword{datasets}
