% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto_get_formatted_palette.R
\name{cart_get_formatted_palette}
\alias{cart_get_formatted_palette}
\title{get one or more color palettes from Carto}
\usage{
cart_get_formatted_palette(carto_palette = NULL)
}
\description{
get one or more color palettes from Carto
}
