% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_get_global_data.R
\name{co2_get_global_data}
\alias{co2_get_global_data}
\title{Get Global CO2 Data}
\usage{
co2_get_global_data(
  paths = list(path_daily_data =
    "https://gml.noaa.gov/webdata/ccgg/trends/co2/co2_trend_gl.csv", path_monthly_data =
    "https://gml.noaa.gov/webdata/ccgg/trends/co2/co2_mm_gl.csv", path_yearly_data =
    "https://gml.noaa.gov/webdata/ccgg/trends/co2/co2_annmean_gl.csv")
)
}
\arguments{
\item{paths}{Paths to the daily, monthly and yearly data}
}
\description{
Get Global CO2 Data
}
