% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_get_keeling_data.R
\name{co2_get_keeling_data}
\alias{co2_get_keeling_data}
\title{Get Keeling Curve Data}
\usage{
co2_get_keeling_data(
  paths = list(path_daily_data =
    "https://gml.noaa.gov/webdata/ccgg/trends/co2/co2_daily_mlo.csv", path_monthly_data =
    "https://gml.noaa.gov/webdata/ccgg/trends/co2/co2_mm_mlo.csv", path_yearly_data =
    "https://gml.noaa.gov/webdata/ccgg/trends/co2/co2_annmean_mlo.csv")
)
}
\arguments{
\item{paths}{Paths to the daily, monthly and yearly data}
}
\description{
Get Keeling Curve Data
}
