% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quarto_starter.R
\name{create_quarto_starter}
\alias{create_quarto_starter}
\title{Createa a quarto document for that project}
\usage{
create_quarto_starter(dir = here(), title = "test", filename = "eda")
}
\arguments{
\item{dir}{The directory where the \code{quarto} folder should be created}

\item{title}{The Title in the quarto document}

\item{filename}{The filename}
}
\value{
nothing. Creates just some files
}
\description{
Createa a quarto document for that project
}
\examples{
create_quarto_starter(title = "Wiener Linien Störungen", filename="wienerLinienEda")
}
