% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections_make_tooltip.R
\name{election_make_tooltips}
\alias{election_make_tooltips}
\title{Make Tooltip for DW Chart}
\usage{
election_make_tooltips(
  df,
  color_mapping = list(list(colname = "ÖVP", color = "#62c3d0", dw_name = "vp"),
    list(colname = "SPÖ", color = "#ff6363", dw_name = "sp"), list(colname = "FPÖ",
    color = "#0066b3", dw_name = "fp"), list(colname = "GRÜNE", color = "#82ae47",
    dw_name = "grne"), list(colname = "NEOS", color = "#e84188", dw_name = "neos"),
    list(colname = "DNA", color = "#bf0000", dw_name = "dna"), list(colname = "KPÖ",
    color = "#bf0000", dw_name = "kp")),
  barHeight = "10px",
  fontSize = ".7rem",
  bg_color = "#f0eded"
)
}
\arguments{
\item{df}{The dataframe with the percetage (!importannt as the bar widths will be computed using this)}

\item{color_mapping}{A list where each entry is a list containing the information on each column you want in the tooltip. \code{colname} = The name in the \code{df}. \code{color} = The color of the bar
\code{dw_name} = Must match the naming-convention of Datawrapper}

\item{barHeight}{The bar height (a css string)}

\item{fontSize}{The fontsize of the party labels (a css string)}

\item{bg_color}{The bg color of each bar}
}
\description{
Creates the tooltip-code for a vertical bar chart. Importantly!! The 'colname'-key in each list entry must match the column-name of the dataframe which holds the percentages.
The 'dw_name'-entry muss be how Datawrapper would write this column. Mostly lower case and connected with lowerstring (_).
}
