% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections_get_winning_party.R
\name{elections_get_winning_party_wide}
\alias{elections_get_winning_party_wide}
\title{Get the winning party and value}
\usage{
elections_get_winning_party_wide(
  df,
  cols,
  col_name_winning_party = "winning_party",
  col_name_winning_val = "winning_val"
)
}
\arguments{
\item{df}{The dataframe}

\item{cols}{a character-vector naming the columns which hold the values}

\item{col_name_winning_party}{the name of the new column containing the name of the winning party/candidate}

\item{col_name_winning_val}{the name of the new column containing the value of the winning party/candidate}
}
\description{
From a table with one (gemeinde, sprengel...) per line. Each column should be named after one candidate/party
}
