% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosphere_get_data.R
\name{geosphere_get_data}
\alias{geosphere_get_data}
\title{Get geosphere data}
\usage{
geosphere_get_data(
  ...,
  api_url = "https://dataset.api.hub.geosphere.at",
  version = "v1",
  type = "grid",
  mode = "historical",
  resource_id = "spartacus-v2-1d-1km",
  output_file = tempfile()
)
}
\arguments{
\item{output_file}{}
}
\description{
Get geosphere data
}
