% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosphere_get_datasets.R
\name{geosphere_get_datasets}
\alias{geosphere_get_datasets}
\title{Get all the datasets form Geosphere}
\usage{
geosphere_get_datasets(
  url = "https://dataset.api.hub.geosphere.at/v1/datasets",
  html = F
)
}
\arguments{
\item{url}{URL to the datasets. Currently \code\link{https://dataset.api.hub.geosphere.at/v1/datasets}}
}
\value{
a dataframe with all the info to the datasets
}
\description{
Basically the same information as on the \link\href{https://dataset.api.hub.geosphere.at/v1/docs/}{ressource-page}
}
