% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosphere_get_prediction.R
\name{geosphere_get_prediction}
\alias{geosphere_get_prediction}
\title{Weather Forecast}
\usage{
geosphere_get_prediction(
  host,
  x,
  y,
  level = NULL,
  subdataset = "T",
  download_dir = "data_raw",
  timezone = "Europe/Paris"
)
}
\arguments{
\item{host}{Character. The data source to query. Must be one of:
\describe{
\item{\code{"nwp"}}{Numerical Weather Prediction (NWP) model outputs. These are high-resolution forecasts for the extended Alpine region on a 2.5 km grid, updated every 3 hours with a forecast horizon of 60 hours. For more information, visit \url{https://data.hub.geosphere.at/dataset/nwp-v1-1h-2500m}.}
\item{\code{"nowcast"}}{Short-term forecasts using real-time observations and radar data, useful for rapid updates. For more information, visit \url{https://data.hub.geosphere.at/dataset/nowcast-v1-15min-1km}.}
\item{\code{"ensemble"}}{The C-LAEF (Convection-permitting Limited-Area Ensemble Forecasting) system provides weather forecasts for the extended Alpine region on a 2.5 km grid. It is updated twice daily with a forecast horizon of 60 hours. The dataset includes the 10th, 50th, and 90th percentiles from 17 model runs (16 ensemble members + 1 control run), allowing for uncertainty estimation in the forecasts. For more information, visit \url{https://data.hub.geosphere.at/dataset/ensemble-v1-1h-2500m}.}
}}

\item{x}{aaa}
}
\description{
Get weather predictions for Austria/Alps
}
