% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosphere_download_stations.R
\name{geosphere_get_stations}
\alias{geosphere_get_stations}
\title{Download all the avaiable stations for the ZAMG daily products}
\usage{
geosphere_get_stations(
  url = "https://dataset.api.hub.zamg.ac.at/v1/station/historical/klima-v2-1d/metadata",
  outputPath = NULL
)
}
\description{
Download all the avaiable stations for the ZAMG daily products
}
