% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosphere_stations_in_lhs.R
\docType{data}
\name{geosphere_stations_in_lhs}
\alias{geosphere_stations_in_lhs}
\title{A dataset with all active stations in the 9 Landeshauptstädte in Austria}
\format{
An object of class \code{sf} (inherits from \code{grouped_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 12 columns.
}
\usage{
geosphere_stations_in_lhs
}
\description{
A dataset with all active stations in the 9 Landeshauptstädte in Austria
}
\keyword{datasets}
