% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_colors.R
\name{get_linear_colors}
\alias{get_linear_colors}
\title{Get Linearly interpolated color from Array of colors}
\usage{
get_linear_colors(df, col, colors)
}
\arguments{
\item{df}{a dataframe}

\item{col}{a numeric column}

\item{colors}{a vector of colors}
}
\description{
Get Linearly interpolated color from Array of colors
}
