% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadZamgStations.R
\name{getZamgStations}
\alias{getZamgStations}
\title{Download all the avaiable stations for the ZAMG daily products}
\usage{
getZamgStations(
  url = "https://dataset.api.hub.zamg.ac.at/v1/station/historical/klima-v1-1d/metadata",
  outputPath = NULL
)
}
\description{
Download all the avaiable stations for the ZAMG daily products
}
