% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZamgVariables.R
\name{getZamgVariables}
\alias{getZamgVariables}
\title{Get ZAMG Variables}
\usage{
getZamgVariables(
  url = "https://dataset.api.hub.zamg.ac.at/v1/station/historical/klima-v1-1d/metadata",
  outputPath = NULL
)
}
\description{
Get ZAMG Variables
}
