% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landeshauptstaedte.R
\docType{data}
\name{landeshauptstaedte_geo}
\alias{landeshauptstaedte_geo}
\title{Geodaten der 9 Landeshauptstädte}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.
}
\usage{
landeshauptstaedte_geo
}
\description{
A dataset of type \code{sf} that contains the polygons for the 9 Landeshauptsädte in Austria in EPSG: 31287 (MGI / Austria Lambert)
}
\keyword{datasets}
