% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landeshauptstaedte.R
\docType{data}
\name{landeshauptstaedte_gkz}
\alias{landeshauptstaedte_gkz}
\title{gkzs der 9 Landeshauptstädte}
\format{
An object of class \code{character} of length 9.
}
\usage{
landeshauptstaedte_gkz
}
\description{
A dataset of type \code{character} that contains the gkzs for the 9 Landeshauptsädte in Austria
}
\keyword{datasets}
