% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDepth.R
\name{list.depth}
\alias{list.depth}
\title{Determine the max depth of a list in R}
\usage{
list.depth(this, thisdepth = 0)
}
\arguments{
\item{this}{a list}

\item{thisdepth}{default depth if \code{this} is not a list}
}
\description{
Determine the max depth of a list in R
}
