% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_gem_data.R
\name{oe_gem_data}
\alias{oe_gem_data}
\title{get the gemeinde geodata for a specific year}
\usage{
oe_gem_data(
  year = 2023,
  base_url = "https://data.statistik.gv.at/data/OGDEXT_GEM_1_STATISTIK_AUSTRIA_",
  dest_file = NULL,
  dest_dir = tempdir(),
  wien_union = F,
  force = F
)
}
\arguments{
\item{year}{}

\item{base_url}{url to where the website with all the years is}

\item{dest_file}{the file to where you want to download the zipped gemeinden}

\item{dest_dir}{folder where the unzipped data will be}
}
\description{
get the gemeinde geodata for a specific year
}
