% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_pop_data.R
\name{oe_pop_data}
\alias{oe_pop_data}
\title{get population data for Austrian gemeinden}
\usage{
oe_pop_data(
  url =
    "https://www.statistik.at/fileadmin/pages/405/Bev_zu_Jahresbeginn_Gebietseinheiten_Zeitreihe.ods"
)
}
\description{
get population data for Austrian gemeinden
}
