% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_get_gleichstellung.R
\name{sora_get_gleichstellung}
\alias{sora_get_gleichstellung}
\title{Get the gleichstellungsindex values}
\usage{
sora_get_gleichstellung(
  url =
    "https://www.sora.at/fileadmin/downloads/projekte/2022_SORA_Datensatz_Staedtebund_Gleichstellungsindex_2021_Dimensionsindizes.xlsx"
)
}
\arguments{
\item{url}{Url to the xlsx file at the sora page}
}
\description{
Get the gleichstellungsindex values
}
