% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistik_utils.R
\name{statistik_format_raw_lsv}
\alias{statistik_format_raw_lsv}
\title{Format raw lst data sent by the statistik austria}
\usage{
statistik_format_raw_lsv(data_year)
}
\description{
Format raw lst data sent by the statistik austria
}
