% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistik_get_pendler_data.R
\name{statistik_get_pendler_data}
\alias{statistik_get_pendler_data}
\title{get the pendler saldo for a gemeinde to all other gemeinden}
\usage{
statistik_get_pendler_data(gkz, year = 2020, no_binnenpendler = T)
}
\description{
get the pendler saldo for a gemeinde to all other gemeinden
}
