% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistik_get_urban_rural_typo.R
\name{statistik_get_urban_rural_typo}
\alias{statistik_get_urban_rural_typo}
\title{get the urban rural typo}
\usage{
statistik_get_urban_rural_typo(source = "eu")
}
\arguments{
\item{source}{either "statistik" or "eu"}
}
\description{
get the urban rural typo
}
