% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistik_utils.R
\name{statistik_prepare_migration_data}
\alias{statistik_prepare_migration_data}
\title{donwload and prepare migration data}
\usage{
statistik_prepare_migration_data(url, gkz_id)
}
\description{
donwload and prepare migration data
}
