% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistik_lst_data.R
\name{statistik_read_lst}
\alias{statistik_read_lst}
\title{Process the statistik lst data}
\usage{
statistik_read_lst(
  dir_with_excels = "~/projects/dst/daten/DATEN_LST_STATISTIK/Standard (Robin Kohrs)/",
  sex = "Zusammen",
  type = "Nettobezüge",
  variable = "Median"
)
}
\arguments{
\item{sex}{Frauen, Männer, Zusammen}

\item{type}{Bruttobezüge, Nettobezüge, SV-Beiträge, einbehaltene Lohnsteuer}

\item{variable}{Anzahl, Median, Mittelwert, P10, P20, P30, P40, P60, P70, P80, P90, Summe}
}
\description{
Process the statistik lst data
}
