% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamg_cleanStationResponse.R
\name{zamg_cleanStationData}
\alias{zamg_cleanStationData}
\title{Prepare the data from a call to the station data}
\usage{
zamg_cleanStationData(data)
}
\description{
Prepare the data from a call to the station data
}
