% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamg_download1x1.R
\name{zamg_get_1x1}
\alias{zamg_get_1x1}
\title{Download Zamg 1 by 1 km Gridded Spartacus data}
\usage{
zamg_get_1x1(
  format = "netcdf",
  parameter = NULL,
  start_date = as.Date("2023-03-20"),
  end_date = Sys.Date(),
  bb = NULL,
  outputFilename = NULL
)
}
\arguments{
\item{parameter}{Either "Tx, Tm, RR"}

\item{start_date}{start date in yyyy-mm-dd}

\item{end_date}{end date in yyyy-mm-dd}

\item{bb}{a bounding box. Either a \code{sf-object} (then the bounding box of the sf object will be computed),
or a \code{bbox-object}}

\item{outputFilename}{The output file name (must have the ending \code{.nc})}
}
\description{
More info \link[here]{https://dataset.api.hub.zamg.ac.at/v1/docs/quickstart.html}.
Test the API \link[here]{https://dataset.api.hub.zamg.ac.at/v1/openapi-docs#/grid/Current_Grid_Data_grid_current__resource_id__get}
}
