% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamg_getPointData.R
\name{zamg_get_point_data}
\alias{zamg_get_point_data}
\title{Get data from Zamg Station data}
\usage{
zamg_get_point_data(
  dataset = "1d",
  stationId = NULL,
  parameters = NULL,
  start = NULL,
  end = NULL
)
}
\description{
Get data from Zamg Station data
}
