% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamg_downloadStations.R
\name{zamg_get_Stations}
\alias{zamg_get_Stations}
\title{Download all the avaiable stations for the ZAMG daily products}
\usage{
zamg_get_Stations(
  url = "https://dataset.api.hub.zamg.ac.at/v1/station/historical/klima-v1-1d/metadata",
  outputPath = NULL
)
}
\description{
Download all the avaiable stations for the ZAMG daily products
}
