% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamg_getPointData.R
\name{zamg_getPointData}
\alias{zamg_getPointData}
\title{Get data from Zamg Station data}
\usage{
zamg_getPointData(
  dataset = "1d",
  stationId = NULL,
  parameter = NULL,
  start = NULL,
  end = NULL
)
}
\description{
Get data from Zamg Station data
}
