% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamg_getStationVariables.R
\name{zamg_getStationVariables}
\alias{zamg_getStationVariables}
\title{Get ZAMG Variables}
\usage{
zamg_getStationVariables(
  url = "https://dataset.api.hub.zamg.ac.at/v1/station/historical/klima-v1-1d/metadata",
  outputPath = NULL
)
}
\description{
Get ZAMG Variables
}
